更新資料型態減少記憶體使用

+++ kernels/kernel_program.cl
@@ -1,5 +1,5 @@
 __kernel void sum_beton_total_amount(
-         __global const int* mask,
+         __global const ushort* mask,
          __global const float* amount,
          __global float* result,
          const int row_size)
@@ -32,7 +32,7 @@ __kernel void sum_beton_total_amount(
 __kernel void calc_numbers_risk(
          __global const float* total_amount_vector,      /* 各beton加總陣列（本金）*/
          __global const float* total_amount_odds_vector, /* 各beton加總陣列（本金*賠率 - 本金）*/
-         __global const int* answers_matrix,
+         __global const short* answers_matrix,
          __global float* result,
          const int beton_length)
 {
